rename python generator sources
