+++ cmake/libprotoc.cmake
@@ -90,9 +90,9 @@ set(libprotoc_files
   ${protobuf_source_dir}/src/google/protobuf/compiler/php/php_generator.cc
   ${protobuf_source_dir}/src/google/protobuf/compiler/plugin.cc
   ${protobuf_source_dir}/src/google/protobuf/compiler/plugin.pb.cc
-  ${protobuf_source_dir}/src/google/protobuf/compiler/python/python_generator.cc
-  ${protobuf_source_dir}/src/google/protobuf/compiler/python/python_helpers.cc
-  ${protobuf_source_dir}/src/google/protobuf/compiler/python/python_pyi_generator.cc
+  ${protobuf_source_dir}/src/google/protobuf/compiler/python/generator.cc
+  ${protobuf_source_dir}/src/google/protobuf/compiler/python/helpers.cc
+  ${protobuf_source_dir}/src/google/protobuf/compiler/python/pyi_generator.cc
   ${protobuf_source_dir}/src/google/protobuf/compiler/ruby/ruby_generator.cc
   ${protobuf_source_dir}/src/google/protobuf/compiler/subprocess.cc
   ${protobuf_source_dir}/src/google/protobuf/compiler/zip_writer.cc
@@ -118,8 +118,8 @@ set(libprotoc_headers
   ${protobuf_source_dir}/src/google/protobuf/compiler/objectivec/objectivec_helpers.h
   ${protobuf_source_dir}/src/google/protobuf/compiler/php/php_generator.h
   ${protobuf_source_dir}/src/google/protobuf/compiler/plugin.h
-  ${protobuf_source_dir}/src/google/protobuf/compiler/python/python_generator.h
-  ${protobuf_source_dir}/src/google/protobuf/compiler/python/python_pyi_generator.h
+  ${protobuf_source_dir}/src/google/protobuf/compiler/python/generator.h
+  ${protobuf_source_dir}/src/google/protobuf/compiler/python/pyi_generator.h
   ${protobuf_source_dir}/src/google/protobuf/compiler/ruby/ruby_generator.h
 )
