+++ src/google/protobuf/compiler/python/helpers.h
