+++ src/google/protobuf/compiler/python/pyi_generator.h
